Update __init__.py
The 'greetings' property doesn't exist, and prevents the python package from running, declaring that it cant reference greetings from the main file (because its not there). Rather than add a greeting, the change made here allows the program to refer to the main file without halting it because of an undefined property. Tested to work on Linux and Windows 3.10